Fix build pipeline for documentation
The doc page shows the git version on top, however, the CI script did
not pull the tags, hence an invalid "untagged" version was shown.
On top, the version and release variables are now set to the expected
X.Y and X.Y.Z format, respectively. In particular, this suppresses
+DISTANCE.gHEX[.dirty] suffixes.

Further, the docs/requirements.txt is not needed as its content was
already reflected in setup.py.