Make testing of tod more flexible
The naming of the parameter `tod` was suboptimal. In `random_type2_message()` it was used as a boolean, yet it was easy to misread the API and pass a number instead. Now, the message tod is set to a random number if `tod` is set to "random". Otherwise, the value of `tod` (a float or int) is used for the message.